Fix client SSL connections from Portus.
Setting the `SSL_CERT_DIR` environment variable means that the standard
certificate paths are ignored, so when Portus tries to connect to a registry as
a client (e.g. when a delete operation is requested) over SSL, it generates an
SSL verify error unless the ca is present in the `/certs` dir - the problem is
standard paths like `/etc/ssl/certs` are ignored, and you get verify errors from
trusted SSL certificates.